Interfaces til OC3 samt ændret Domain-package til lille "d"
