+++ src/logic/Dimensionerendekraft.java
@@ -1,0 +1,9 @@
+package logic;
+
+public interface Dimensionerendekraft {
+
+	public double getNewton();
+	public static final double tyndekraft = 9.816;
+	
+	
+}
