+++ src/logic/DimensionerendekraftEjDefineretException.java
@@ -1,0 +1,5 @@
+package logic;
+
+public class DimensionerendekraftEjDefineretException extends Exception {
+
+}
