+++ src/logic/NormalkraftEjDefineretException.java
@@ -1,0 +1,5 @@
+package logic;
+
+public class NormalkraftEjDefineretException extends Exception {
+
+}
