+++ src/logic/PTECalculatorController.java
@@ -1,0 +1,10 @@
+package logic;
+
+public interface PTECalculatorController {
+
+	public void beregnNormalkraft();
+	
+	public double getNormalkraft();
+
+	
+}
