+++ src/logic/PTEObserver.java
@@ -1,0 +1,6 @@
+package logic;
+
+public interface PTEObserver {
+
+
+}
