+++ src/logic/Vinkel.java
@@ -1,0 +1,8 @@
+package logic;
+
+public interface Vinkel {
+	
+	public double getGrader();
+	
+	public boolean getMaaltTilLodret();
+}
