+++ src/logic/VinkelEjDefineretException.java
@@ -1,0 +1,5 @@
+package logic;
+
+public class VinkelEjDefineretException extends Exception {
+
+}
